refactor(cli): move chunk management to its own module
