+++ sn_cli/src/subcommands/wallet.rs
@@ -19,9 +19,7 @@ use std::{
     path::{Path, PathBuf},
     str::FromStr,
 };
-use tempfile::tempdir;
 use url::Url;
-use xor_name::XorName;
 
 use super::files::chunk_path;
 
@@ -150,27 +148,21 @@ pub(crate) async fn wallet_cmds_without_client(cmds: &WalletCmds, root_dir: &Pat
 pub(crate) async fn wallet_cmds(
     cmds: WalletCmds,
     client: &Client,
-    wallet_dir_path: &Path,
+    root_dir: &Path,
     verify_store: bool,
 ) -> Result<()> {
     match cmds {
-        WalletCmds::Send { amount, to } => {
-            send(amount, to, client, wallet_dir_path, verify_store).await?
-        }
-        WalletCmds::Receive { file, transfer } => {
-            receive(transfer, file, client, wallet_dir_path).await?
-        }
-        WalletCmds::GetFaucet { url } => get_faucet(wallet_dir_path, client, url.clone()).await?,
+        WalletCmds::Send { amount, to } => send(amount, to, client, root_dir, verify_store).await?,
+        WalletCmds::Receive { file, transfer } => receive(transfer, file, client, root_dir).await?,
+        WalletCmds::GetFaucet { url } => get_faucet(root_dir, client, url.clone()).await?,
         WalletCmds::Pay {
             path,
             batch_size: _,
         } => {
-            let file_api: Files = Files::new(client.clone(), wallet_dir_path.to_path_buf());
+            let file_api: Files = Files::new(client.clone(), root_dir.to_path_buf());
 
-            // Temp folder to hold SE chunks, which is cleaned up automatically once out of scope.
-            let temp_dir = tempdir()?;
-
-            let chunked_files = chunk_path(&file_api, &path, temp_dir.path()).await?;
+            let chunk_artifacts = root_dir.join("chunk_artifacts");
+            let chunked_files = chunk_path(&file_api, &path, &chunk_artifacts).await?;
 
             let all_chunks: Vec<_> = chunked_files
                 .values()
@@ -366,9 +358,4 @@ async fn receive(transfer: String, is_file: bool, client: &Client, root_dir: &Pa
 
     println!("Successfully stored cash_note to wallet dir. \nOld balance: {old_balance}\nNew balance: {new_balance}");
     Ok(())
-}
-
-pub(super) struct ChunkedFile {
-    pub file_name: String,
-    pub chunks: Vec<(XorName, PathBuf)>,
 }
